updates delete() to check that the submitted username matches the logged in user

+++ src/app.py
@@ -47,27 +47,32 @@ def delete():
     if 'loggedin' in session:
         if request.method == 'POST' and 'username' in request.form and 'monthlybudget' in request.form:
             username = request.form['username']
-            budget = request.form['monthlybudget']
+
+            if username != session['username']:
+                msg = 'Incorrect username.'
             
-            cursor = cnx.cursor(buffered = True)
-            query = "SELECT username, monthly_budget FROM user WHERE username = %s"
-            cursor.execute(query, (username,))
-            account = cursor.fetchone()
-            cursor.close()
-            if account:
-                # confirm monthly budget
-                if float(budget) == float(account[1]):
-                    deletecursor = cnx.cursor(buffered = True)
-                    query = "DELETE FROM user WHERE username = %s and monthly_budget = %s"
-                    deletecursor.execute(query, (username, budget,))
-                    cnx.commit()
-                    msg = 'Account successfully deleted.'
-                    deletecursor.close()
-                    return redirect(url_for('login'))
+            else: 
+                budget = request.form['monthlybudget']
+                
+                cursor = cnx.cursor(buffered = True)
+                query = "SELECT username, monthly_budget FROM user WHERE username = %s"
+                cursor.execute(query, (username,))
+                account = cursor.fetchone()
+                cursor.close()
+                if account:
+                    # confirm monthly budget
+                    if float(budget) == float(account[1]):
+                        deletecursor = cnx.cursor(buffered = True)
+                        query = "DELETE FROM user WHERE username = %s and monthly_budget = %s"
+                        deletecursor.execute(query, (username, budget,))
+                        cnx.commit()
+                        msg = 'Account successfully deleted.'
+                        deletecursor.close()
+                        return redirect(url_for('login'))
+                    else:
+                        msg = 'Incorrect monthly budget.'
                 else:
-                    msg = 'Incorrect monthly budget.'
-            else:
-                msg = 'Username does not exist.'
+                    msg = 'Username does not exist.'
         
         elif request.method == 'POST':
             msg = 'Please fill in the empty fields.'
